search: extract SearchContext
This is a mechanical refactor that extracts read-only fields from
searchResolver and exports them. We also update the interface
`searchResolver` so that we can access the SearchContext from the
streaming consumer.

The purpose of this refactor is to make meta fields like `query` and
`patternType` accessible for alert creation in streaming.

The naming is obviously terrible, suggestions are welcome!