search: extract FileMatch from FileMatchResolver (#17706)
This is a first step towards a FileMatch type which doesn't depend on
GraphQL resolvers. It still depends on GitCommitResolver via the symbol
results. However, this commit detaches the dependency on
RepositoryResolver.

Our goal is to extract core search logic out of the graphqlbackend. Our
next commit will replace uses of FileMatchResolver with FileMatch. These
will only be converted into FileMatchResolvers in the
SearchResultsResolver.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>