search: extract SearchInputs (#17752)
This is a mechanical refactor that extracts read-only fields from
searchResolver and exports them. We also update the interface
`searchResolver` so that we can access the SearchInpus from the
streaming consumer.

The purpose of this refactor is to make meta fields like `query` and
`patternType` accessible for alert creation in streaming.