search: periodically send ping event when streaming (#17766)
In long running searches with not many results we have many seconds go
by with no events sent. This may lead to something in between (proxy or
HTTP server?) timing out the connection between the browser and our
backend. So we introduce a periodic ping event. We make it useful by
just sending down the current state of progress.

Co-authored-by: Erik Seliger <erikseliger@me.com>