search: only set progress.repositories_count at end
From testing streaming search it was confusing that the repositories
count would be set while the stream was going. It is confusing since a
user would assume this is the number repositories that have been
searched, rather than the total number that _will_ be searched. So
instead we only communicate this number at the end.