search: propose count when limit hit (#17864)
In streaming, we don't have a nice way to educate the user to increase
the limit when they hit them. This adds a proposed query for increasing
the count.

To achieve this we needed to know what the current limit is. This was
hidden as an implementation detail of SearchResults. It is now promoted
to the field Limit in SearchInputs.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>