Check for branch duplicates after creating changeset specs (#448)
* Check for branch duplicates after creating changeset specs

This is a follow-up to #442 and ensures that changeset specs are not
getting silently lost by validating that multiple changeset specs in the
same repository have different branches.

I decided to make this a separate step _after_ the execution of the
steps so that users can leverage the cache. That allows them to change
the campaign spec and then rerun the command after they get this error,
vs. the execution being aborted after running into this error (if we'd
do the check inside executor).

* Fix naming in duplicateBranchesErr