search: remove error from SearchEvent (#17829)
We expected once a search stream returned an error event, it would
stop. Additionally we expected once a function returned that took a
stream, it would stop writing to the stream. This assumption could lead
to bugs if broken. By removing error from the event, we more clearly
force the above.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>