servegit: eval symlinks on root (#454)
This is a useful use case which allows changing your root by updating a
symlink. Additionally it is convenient. We have to explicitly resolve
the symlink since filepath.Walk does not support resolving symlinks. We
resolve the symlink each time we crawl so the user can update the value
of the symlink at runtime. Note: we still ignore symlinks with the repos
directory. Supporting that would be more complex due to symlink cycles.