search: factor out most functionality in searchFilesInRepos (#17952)
This function has been confusing to use due to the size of it, and quite
a few special cases around structural and searcher. This PR pulls out
most of the logic into separate functions. This leads to a much easier
to navigation function which is mainly responsible for calling out to
zoekt/searcher/etc.

One of the main things done here is a switch to using errgroup
more.