search: return early if time left for zoekt is <=0
It was possible that we start calling zoekt with a negative wall time
left on the clock. This didn't cause any issues, because the zoekt
client would return early. However, by moving repo resolution before the
first call to zoekt, we waited endlessly if the deadline had already
exceed because the context we hand to getRepo never expires.

This little change makes sure we have at least some time left on the
clock before we call zoekt.

In a follow-up PR I will move repo resolution to before this block.