search: unify code for missing status on unindexed repos (#17980)
We had duplicated code between symbols and text search. Additionally it
was tricky in searchFilesInRepos since you had to use the searcherRepos
variable, not the "indexed.Unindexed" variable. Now there is only
indexed.Unindexed.

Previously we would log to the tracer about missing vs not. This has
been removed. The information is not that valuable these days.