search: introduce startSearch function
We combine newSearchResolver and newResultsStream into a single
function. This abstracts away what we need to do to setup the search
stream. Additionally we do not expose the "finalC" channel, and instead
return a function. This is cleaner for the client.

This commit was motivated by making the next commit more minimal.