git: use custom format for git log (#17983)
Git has supported %S since 2.21 (Q1 2019). Git commit search was written
before that at Sourcegraph so we had a workaround using "--source". Now
that it is supported we can more directly use it instead of ignoring
extra information on the line.

Additionally we may be able to drastically simplify commit search and
just call log. A todo is left in the code for future travellers.