search: apply limit across backends in streaming
Now that streaming has higher defaults we can apply the limit across
backends. Previously the default limit was 30 so we didn't want one
search type to dominate results. Now that the limit is 500 we can be
more aggressive with cancellation. This is an effort to increase
confidence in result counts.