insights: add initial DB schema
This change adds the initial code insights DB schema. It is generally well-thought-out
and I did a fair amount of analysis on whether or not to use more involved TimescaleDB
features like [compression](https://github.com/sourcegraph/sourcegraph/pull/17227#issuecomment-773704129)
instead of / in addition to storing repo names as IDs, for example.

I am sure we will change the schema as we learn more, but this generally seems
sufficient/appropriate for now.

Helps #17218

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>