db: add a dev script to renumber migrations
When I'm working on feature branches that take more than a day or two to
finish, review, and merge, I find that my migrations often end up
conflicting with other people's migrations that have been merged in the
interim. This adds instructions on how to deal with this, based on what
I personally do, and (probably more usefully) bundles the core of the
process into a script that can do most of the dirty work.