Normalize changeset spec fields to speed up joins (#18110)
This is the result of @eseliger and me running into really slow
`GetRewirerMapping` queries yesteday.

We found out that a single `changeset_spec` with an abnormally large
`spec` column was the cause of this performance regression:

1. The `spec->'diff'` contained ~57mil lines of code.
2. The previous `GetRewirerMapping` query caused the database to
   apparently load the complete JSONB columns from disk.

57 million lines in a single diff is not normal, but it was the result
of accidentally adding the `node_modules` folder to the diff. And *that*
is not that uncommon.

And such a single, accidentally-uploaded changeset spec caused
performance problems for the complete instance.

So we decided to normalise the 3 `spec` fields that we use for JOINs and
WHERE conditions by putting them in separate columns, with indexes.

No performance problems anymore and the queries got quite a bit cleaner
too.

**Note** though that these columns are **write-only** for now and are
only read by SQL queries. We didn't see a need to make use of them in
application code for now, but that may change in the future at which
point we should add them to the `ChangesetSpec` type definition.