schemadoc: Render view query in schema documentation (#18280)
I find it very hard to see what changed when someone has a migration that recreates a view, because the git diff will just be all new lines. This implements adding the backing query to the generated schema file, which should make it easier to discover what actually changed in a migration that touches views.