search: use interface assertion on select
This is a minor change. I noticed when reading the code that we did
explicit type assertions for each result type. However, the function we
called in each case matched the same interface. Instead we can just
assert on the interface.

If we want to enforce that each result type matches the interface we can
either add it to the SearchResultResolver interface or we can add type
assertions for each result type.