gitserver: set Content-Type and Cache-Control on exec
We hadn't done this before. Our client doesn't care about this, so it
shouldnt' affect anything. The only possibility it may affect is direct
callers to `git archive`. However, this content type is more correct
since it actually indicates the data can be binary. So this would likely
lead to better behaviour in other cases.