dev: allow codeinsights-db more time to startup
I notice in some cases codeinsights-db needs some more time to start because
my CPU is consumed by e.g. webpack or similar and building the Docker image
takes a bit longer.

I think we can actually make relying on the codeinsights-db on startup optional
(it should be, with errors just communicated at query-time, I think) - I will
send a PR to do this but for now this is better.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>