insights: turn on for Sourcegraph.com (#18360)
This PR turns on insights code Sourcegraph.com only (not customer deployments,
not our dogfood deployment, etc.)

Rollout should go smoothly, but in case it doesn't there is an escape hatch
`DISABLE_CODE_INSIGHTS=true` we can set to disable it. Only reason not to turn
on for all customers currently is because the PR for TimescaleDB in Docker Compose
hasn't yet been sent - that will come in the next day or so.