gitserver: set transfer-encoding to chunked for exec (#18321)
We have had a "flushingResponseWriter" since the dawn of time. However,
I believe it has never been effective since we never set the
transfer-encoding to chunked. So the server may be flushing data
constantly, but our clients would not be informed that data would often
be flushed.

This was originally added to make things like diff/commit search faster
if I remember. This now correctly does that. For example on
sourcegraph.com if you do a streaming search you will now see results
much faster.