Search context: Add search context url param and append context to queries (#18209)
Adds `context` param to the URL which corresponds to the selected search context in the dropdown. If a context filter is not present in the query, then the selected search context gets appended to the query.

When switching between search contexts we automatically re-run the search (similar behaviour as the version contexts dropdown).