extensions: stop ref counting for extensions query (#18149)
On webapp initialization, we fetch enabled extensions whenever contributions are registered, so this results in network console spam. `viewerConfiguredExtensions` was multicast before, but since we counted references (subscribers), it would repeatedly cancel and restart the query since the query hadn't completed yet and couldn't be buffered. Remove `publishReplay` + `refCount` in favor of `shareReplay(1)`