insights: store: add fast data existence check
For backfilling of historical data, we need a way to check if data in a given
timeframe does not exist for a series. It doesn't make sense to query the actual
data in that case, and so we add this method to get just the unique series IDs
that do have at least one data point in a specific timeframe.

Also fix an issue where the timeframe was not inclusive as it should've been.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>