Improve performance of VirtualList
while we didn't _display_ all elements at once with the VirtualList, it still required to render the react components in memory, before we pass them in. For searches with a lot of results, this caused thousands of renders to happen on each rerender (ie when streaming search receives a new event). This is fixed by providing an explicit render function, that renders the items lazily, instead of being passed-in the eager rendered components.