Distinguish permission errors from 404 deleted when loading changesets (#18256)
Before, these errors were not handled properly and a token not having the right scope would mark a changeset as deleted in a syncer run.
This fixes it by relying on the fact that every pull/merge request in a repo can be seen, if the repo can be seen. So if the error is a project not found error, it should be safe to assume that there is no access, instead of the changeset being deleted.
This was already working for GitHub, but not for GitLab and Bitbucket Server.
Sadly, GitLabs error reporting is a bit vague, but it works good enough, I think.