search: Split out CommitSearchResult (#18497)
Splits CommitSearchResult from CommitSearchResultResolver so that
graphql-specific functionality can be isolated from the result type, and
the result type can be moved to `internal/search`.

This is just a first step, and it's not yet ready to move out of this
packag. Next is to remove the dependency on the GitCommitResolver, which
makes requests to hydrate its fields.