insights: enable in all production deployments
This change effectively enables code insights for all production deployments.

It is not turned on for single-container deployments (we will not support running
the backend there as those deployments are for demos only.)

It can be disabled by setting `DISABLE_CODE_INSIGHTS=true` on both the frontend
and repo-updater containers; in which case it will not run at all anywhere.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>