insights: ensure search queries run with a large count: (#18518)
This ensures insights without a `count:` field in them have `count:9999999` (10m)
implicitly added. All search queries have a timeout of 60s, so this is unlikely to
cause any major issues. Additionally, is is a source of large issues for code insights
because we implicitly measure "number of fluctuating results" leading to super wonky
things _like this_:

![image](https://user-images.githubusercontent.com/3173176/108571739-15d45b00-72ce-11eb-849b-029032d1ba01.png)

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>