Move Subject and Body to methods on a Message type
Rather than having standalone functions for Subject and Body, we can
move these to methods on a Message type so they can be more easily
re-used and so that they are co-located with the type they operate on.

As a bonus, it removes a reference to graphqlbackend from campaigns resolvers
code.