Hide campaigns in deleted namespaces (#18224)
This fixes breaking the API and hence the global list of campaigns, when a campaign in a deleted namespace exists in the database.
In order to not pick those changesets up in the reconciler, I added a view that filters those out. It also fixes the reconciler picking up changesets in deleted repos.

Note that hiding may not be the final behavior we want here, but if fixes the bug, so it's a good stopgap for now, I think.