search: Untangle commit results (#18525)
This commit includes a series of changes that untangles commit results from commit resolvers. It also significantly simplifies many parts of the commit search result construction since many parts of the search result are really resolver specific and can live in the resolver methods. 

* Move GitCommitResolver out of CommitSearchResult

Removes `GitCommitResolver` from `CommitSearchResult` and replaces it
with a git commit and repo name from which resolvers can be constructed.

* Move creation of ref resolvers into the CommitSearchResultResolver

There is no reason to create the ref resolvers ahead of time, so we can
just save the refs directly on the `CommitSearchResult` and construct
the resolvers when requested.

* Remove redundant string conversion

* Simplify icon

There is no reason to save the icon string literal on the struct, since
this is only used when the resolver's Icon() method is called, and can
just return the string literal directly.

* Remove redundant string conversion

* Add Short() method to api.CommitID

We can move the code to get the short commit ID onto the type so it can
be properly named and documented.

* Move detail to resolver

Detail is only used by the resolver, so we can construct it on request
rather than ahead of time.

* Remove error return from URL

URL does not need to return an error, which simplifies a few things up
the chain.

* Move Subject and Body to methods on a Message type

Rather than having standalone functions for Subject and Body, we can
move these to methods on a Message type so they can be more easily
re-used and so that they are co-located with the type they operate on.

As a bonus, it removes a reference to graphqlbackend from campaigns resolvers
code.

* Create label on the fly

No need to create a label ahead of time since it can be constructed from
other fields when requested.

* Remove unused error

This error is unneeded, so we can remove it to simplify the call chain.

* Remove pre-calculated url from CommitSearchResult

The URL does not need to be constructed in advance, so we can simplify
construction of the CommitSearchResult by moving it to its method call.

* Simplify patString declaration

No need to declare patString ahead of time.

* Simplify construction of CommitSearchResultResolver

Rather than instantiating the CommitSearchResolver, then populating its
fields individually through slice accesses, we can make this
significantly easier to follow by creating the CommitSearchResult on one
go.

* Move resolver creation into CommitSearchResultResolver.Matches()

No need to construct the match ahead of time, so then we don't need to
store resolvers on the result struct.

* Pass in a types.RepoName rather than a RepositoryResolver