search: use O(1) space for computing filters (#18542)
Previously we allocated O(n) space for computing filters. We created a
slice of all filters, sorted then kept the best 24 filters. Instead we
can use our hard earned computer science degrees to use O(1) space with
a MinHeap.

Note: we already allocate O(n) space in the map of all filters. This
commit just avoids an additional allocation and sort on that.