search: Use symbolResolver rather than searchSymbolResult in suggestions (#18593)
We were previously using searchSymbolResults when creating suggestion
resolvers, which required that we also saved a reference to a database
so we could convert it to a symbolResolver. However, there's no reason
we can't just create the suggestion resolver from a symbolResolver
instead, allowing us to skip saving a database on the suggestion
resolver type.