search: Exercise symbol resolver methods in integration tests (#18629)
The integration tests didn't catch a very basic panic condition because
the graphql query didn't actually exercise any of the symbol result
methods. This adds those fields to the query (but does not validate the
output) so that we can at least catch very high level issues with symbol
search.