progress-bot: Improved progress bot
This commit introduces a new version of our progress bot, built in Go,
that knows how to only report the entries in the CHANGELOG file over the
last 24h.

It has a few other bells and whistles, which I had fun putting together,
like building group image avatars of the contributors of each section.

It runs as a GitHub action on a 24h schedule, reporting new entries to
the #progress slack channel, with much nicer formatting than before.