search: Refactor suggestions and add integration tests for them (#18631)
* search: Refactor suggestions

This commit refactors suggestions to be implementors of the
searchSuggestionResolver interface. This makes all the suggestions
well-typed so we aren't casting from an `interface{}` value, allowing
static analysis to be more useful. The fact that
`newSearchSuggestionResolver` took an `interface{}` argument contributed
to a recent outage from a panic further down the line, so I think this
is worth doing even if the change is fairly large and adds some
boilerplate.

* search: Add gqltest for suggestions

Adds framework and a simple integration test for suggestions.