frontend: remove session upon signout (#18686)
This change deletes both the session cookie and the Redis session key when the client logs out.
Currently, when a user logs out, the session is invalidated but not deleted. Whenever a client
logs in with the same browser, the old session is reused by github.com/gorilla/sessions
which can lead to authentication bypass.