Add minimal enforced linting rules (#18703)
* Add minimal enforced linting rules

Adds a minimal `golangci-lint` configuration to enforce linting across
the project. Currently, the only lint it enforces is `ineffassign` since
there were only a couple of instances to change for that.

A separate configuration (`.golangci.enforced.yml`) was added so linting
behavior for people who have it set up as an editor integration does not
change. I think that ideally, we should aim to unify these, but for now
this strikes a nice balance between nagging lints and lints that cause CI
to fail.

* Add comment with link to tracking issue