internal/vcs/git: add FindNearestCommit to get commit nearest to a point in time
Stacked on top of #18712

For building historical data of search insights, we need a way to locate the commit
in a repository closest to some given point in time so that we can search it using
a `repo:<repo>@<rev>` filter and determine the number of results that existed around
that point in time. This adds the ability for us to determine which commit is most
appropriate for a given time range.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>