search: buffer sender in zoektSearch
We add a buffer to the Sender in `graphqlbackend/zoektSearch` to reduce
the backpressure that Sourcegraph excerts on Zoekt.

In our current code we don't buffer which most likey causes Zoekt to
block until repos are resolved.

The buffer size is hard-coded at 30 `zoekt.SearchResults`, which should
be more than enough to fill the first page.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>