dev: build without custom build tags
None of our go files specify the dev or delve build tag. It is possible
a dependency specifies the delve tag, but I don't see any recommendation
in the delve README about this. If it does, we can add back build tag
support for delve. I think we originally added this tag back in the day
for how we shipped delve support (predates OSS).