dev/gqltest: separate search tests which can be used for streaming (#18769)
This is an initial commit which splits the search tests in two. Those
that can be run against the streaming API and those that won't be part
of streaming. Our gql tests ask for very little from the GQL API, so
luckily we can re-use them. The next commit will add a streaming client
as well as run "testSearchClient" twice. Once with the normal GQL client
and once with the streaming client.