Group campaigns graphql schema extension (#18809)
Using the extend keyword, we can make it so all campaigns resolvers are in one place in this file. This should make it easier for us to split it up in three sections as a follow-up:
- Deprecated
- New
- Untouched
and in the future, we can simply cut out the deprecated section. Therefore, we can make use of the fact that extend can be used multiple times.