remove stable:yes
most of the tests specified stable:yes but that switches to paginated
search instead of our normal search path. Can't merge this change in
until we add some tests which test stable:yes again. However, stable
search is not yet supported for streaming search and not sure if it will
be.