search: Make select independent of searchResolver
I'm realizing that some of the difficulty in making changes in this code
is because the *searchResolver receiver on many of these methods is used
and mutated deep down the call tree. We can remove searchResolver as a
receiver on the selectResults function to make its dependencies more
clear and help provide confidence that it is side-effect free.