gitserver: Perform prefix match on hostname
We need to match the gitserver hostname with a value from the
SRC_GIT_SERVERS values which often include a domain and port too.

We could override the HOSTNAME environment but we'd need to remember do
this for customer deployments too so it seems simpler to handle it in
code.