Fix close changesets operation
We set the wrong reconciler_state (.ToDB() lowercases it for workerutil), so it was never picked up by the reconciler.
I think the ultimate fix should be to get rid of the .ToDB call altogether, but it's more effort and I wanted to get this error fixed first, as I don't know if I will get to it before the next release cut.