search: always copy stats.Repos (#18982)
This removes a performance optimization we had around sharing the
underlying map for stats.Repos. We will follow-up with a way to avoid
the extra allocations this change introduces.

The optimization assumed we only created the c.Repos map once. However,
with and/or queries we call doResults per term. Each call to doResults
creates a new c.Repos. This is then aggregated, leading to writes to a
shared c.Repos. Concurrent writes to a map cause a panic.