search: Move highlightedString to internal/search/results
Moves highlightedString out of graphqlbackend in preparation for moving
CommitSearchResult out of graphqlbackend.

Note the somewhat awkward resolver wrappers. This is done because we
need to make the fields of HighlightedString and HighlightedRange
public so the types can be constructed by an external package, but
then those conflict with the graphql resolver methods. The other
option would be to keep the fields private and use constructors for the
types, but I found that even more awkward, and stylewise I prefer to
have "passive" types like our result types to contain all public fields.

Also note the conversion to value types for HighlightedRange. No real
reason to keep these as pointers considering how small they are, and it
makes conversion to resolver types scarier with nil derefs. This will
be made unnecessary by generics...someday.