Fix and enforce bodyclose lint
This commit enables the bodyclose lint as an enforced lint, and fixes
all warnings.

There were a number of lint warnings for test files, but for this
particular lint, it's not very valuable in tests since connection leaks
just don't matter that much, and it would likely just add noise. For
this reason, the lint has been disabled for `_test\.go` files.

There were a few false positives, but it also caught a few cases that
appear to be actual connection leaks, so I think it's still worth enabling
even if it does add a small amount of noise.