Ensure we always use the newest external service (#19033)
I always thought we use the oldest, but the stores default ordering is `id DESC`, so maybe that never was the case.. Looking back git histories a bit, I didn't find evidence it was changed. Regardless, my recent attempt to deprioritize user external services removed that ordering and it was basically random, this fixes it to be stable again, at least. As this logic is going to disappear anyways, I'm not too concerned about it for now.