search: use V2 search API in gqltests (#19057)
We were using the old version of the graphql API for search. This was
the reason some of the streaming search integration tests failed, since
the tests where written against the old version. The old version
defaulted to patternType regexp, while V2 defaults to patternType
literal.

We update the tests to explicitly specify the patternType when
needed. We also update streaming search to only specify patternType as
an argument, if the user specified it. This allows the
"detectPatternType" logic to work correctly if the version is V1 in
streaming.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>