search: Parse predicate params as part of the full value
When parsing predicates like `repo:contains(file:test)`, this would be
parsed like `repo:contains "(file:test)"`. This is done to avoid parsing
subqueries as part of the literal, but for predicate filters, the params
often look a lot like subqueries. This is the minimum change I could
make to make this non-blocking, but it should be revisited quickly for a
less hacky solution.