Rename "default repos" to "indexable repos"
Default repos doesn't really convey the intent anymore. It was
originally just listing repos from the default_repos table on
sourcegraph.com so that they would be indexed. We are now pulling in
repos from more places so calling it "default repos" is confusing. The
intent is still that any repo included in should be indexed.