streaming: Prometheus, track time to first result
This adds a new Prometheus histogram which tracks "time to first result"
for streaming. I chose to create a new histogram because
`graphqlFieldHistogram`, the histogram which currently tracks this
metric for GraphQL-based searches, requires labels, such as `type`,
`field`, `error`, `request_name`, which don't makes sense in the context
of streaming.

In a separate PR we should introduce a new panel on the Grafana
Dashboard "Search API Usage at a Glance" with the histogram as source.