gqltest: detect graphql query name to remove log spam (#19110)
If the name isn't set, the query is logged by Sourcegraph. Given we only
set the name once for a repo and indexed repo query, it resulted in a
lot of log spam. We now extract the name out of the query.

Additionally we remove the intentional logging of the first
WaitForReposToBeCloned and WaitForReposToBeIndex call. This seems low
value and leads to a more complicated API.

The bulk of this commit is removing the first empty string to GraphQL
function calls. See the change to the GraphQL function where we use a
regex to extract the name.