search: log to trace even if ctx.Canceled (#19126)
We don't see the logs we expect in the traces on Cloud although locally
we do. I suspect this has to do with context cancellation after we hit
limits. I believe most, if not all, of the Zoekt instances return with
a canceled context for which we don't log the aggregated stats.

With this change we log zoekt Stats even if StreamSearch returnes with
an error.