search: don't set span to failed for context error
By design, context errors are very common for streaming search. For
example, every time we hit a count limit, `limitStream` cancels the
context. As a result, a lot of spans are marked as failed although we
shouldn't consider them as such. This might be confusing for anyone
trying to troubleshoot issues by investigating traces.

With this change we still log context errors, but we don't set the
corresponding span as failed.